+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+
+## RELEASE 2.5.0 - 2019-10-15
 ### Added
 - Command Generate - Add  "Application port" parameter (`-p, --application-port`) to `lumber generate`.
 - Command Generate - Add  "Application host" parameter (`-H, --application-host`) to `lumber generate`.